+++ R/blocksel.R
@@ -1,24 +1,24 @@
 blocksel <- function(X, blocks) {
   
-  X <- .matrix(X)
-  n <- dim(X)[1]
+    X <- .matrix(X)
+    n <- dim(X)[1]
   
-  nbl <- length(blocks)
+    nbl <- length(blocks)
   
-  selcol <- unlist(blocks)
+    selcol <- unlist(blocks)
   
-  colnam <- colnames(X)[selcol]
+    colnam <- colnames(X)[selcol]
   
-  X <- X[, selcol, drop = FALSE]
-  colnames(X) <- colnam
+    X <- X[, selcol, drop = FALSE]
+    colnames(X) <- colnam
 
-  z <- lapply(1:nbl, function(i) length(blocks[[i]]))
-  lengthblock <- unlist(z)
+    z <- lapply(seq_len(nbl), function(i) length(blocks[[i]]))
+    lengthblock <- unlist(z)
   
-  z <- lapply(1:nbl, function(i) rep(i, lengthblock[i]))
-  newcol <- data.frame(newcol = 1:sum(lengthblock), block = unlist(z))
-  newblocks <- lapply(1:nbl, function(i) newcol$newcol[newcol$block == i])
+    z <- lapply(seq_len(nbl), function(i) rep(i, lengthblock[i]))
+    newcol <- data.frame(newcol = seq_len(sum(lengthblock)), block = unlist(z))
+    newblocks <- lapply(seq_len(nbl), function(i) newcol$newcol[newcol$block == i])
   
-  list(X = X, blocks = newblocks)  
+    list(X = X, blocks = newblocks)  
 
-  }
+    }